predictive_maintenance_agent/requirements.txt
paho-mqtt
jsonschema
scikit-learn
joblib
pandas # pandas is used in the model training script, but also good practice to include if model expects Series/DataFrame
json-logging